+++ src/main/java/org/example/coworking/enums/ResourceType.java
@@ -1,4 +1,4 @@
-package enums;
+package org.example.coworking.enums;
 
 public enum ResourceType {
     WORKSPACE,
