+++ src/main/java/org/example/coworking/enums/UserRole.java
@@ -1,4 +1,4 @@
-package enums;
+package org.example.coworking.enums;
 
 public enum UserRole {
     USER,
